+++ docs/12_0_0/gettingstarted/whatsnew.md
@@ -6,6 +6,8 @@ This page contains a list of big features. Please check the GitHub issues for al
 
 * AutoComplete: added `lazyModel` attribute for fetching autocomplete-suggestions (instead of completeMethod).
 * CommandButton: added `disableOnAjax` attribute to disable the button during Ajax requests triggered by it.
-* CommandLink: now has a widget and can be enabled and disabled by its widget.
+* CommandLink
+  * Now has a widget and can be enabled and disabled by its widget.
+  * Added `disableOnAjax` attribute to disable the button during Ajax requests triggered by it.
 
 Look into [migration guide](https://primefaces.github.io/primefaces/12_0_0/#/../migrationguide/12_0_0) for more enhancements and changes.